supplyiq/requirements.txt
streamlit
langgraph
langchain
langchain-google-genai
google-generativeai
pandas
plotly
python-dotenv
sqlite-utils
openpyxl
